Refactor file paths for `ProductPicker` component and `sales` page
